Add xoshiro256p implementation to ptr_to_vtable.
xoshiro256p is basically a copy of the xoshiro256ss implementation with
one line changed. The difference between them could just be changing the
``struct rand_interface`` object assigned to the ``rand`` field of their
internal states but the copy/pasted separate implementations will be
kept for the purposes of comparing theoretically separate
impelementations of a common interface.